modify player height display

+++ Basket/resources/views/players.blade.php
@@ -24,7 +24,7 @@
             <td> {{ $player->lastname }} </td>
             <td> {{ $player->position }} </td>
             <td> {{ $player->active == 0 ? 'no' : 'yes' }} </td>
-            <td> {{ $player->height/100 . "m" . $player->height%100 }} </td>
+            <td> {{ (int)($player->height/100) . "m" . $player->height%100 }} </td>
             <td> {{ $player->weight }} kg </td>
             <td> {{ $player->gender }} </td>
             <td> {{ date_diff(date_create($player->birthdate), date_create('now'))->y }} </td>
